+++ common-tool-spring-boot-starter/src/main/java/io/github/kk01001/util/NetworkUtil.java
@@ -1,6 +1,7 @@
 package io.github.kk01001.util;
 
 import cn.hutool.core.net.NetUtil;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.util.StringUtils;
 
 /**
@@ -8,6 +9,8 @@ import org.springframework.util.StringUtils;
  * date:  2023-08-07 17:28
  */
 public class NetworkUtil {
+
+    public static final String LOCAL_SERVER_IP = getLocalIp();
 
     private static final String SERVER_IP = "SERVER_IP";
 
@@ -31,5 +34,26 @@ public class NetworkUtil {
             return value;
         }
         return System.getProperty(key);
+    }
+
+    public static String getClientIp(HttpServletRequest request) {
+        // 优先尝试获取 X-Forwarded-For 头中的信息
+        String ip = request.getHeader("X-Forwarded-For");
+
+        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
+            // 如果 X-Forwarded-For 为空，尝试获取 X-Real-IP 头中的信息
+            ip = request.getHeader("X-Real-IP");
+        }
+
+        if (ip == null || ip.isEmpty() || "unknown".equalsIgnoreCase(ip)) {
+            // 如果 X-Real-IP 也为空，最终回退到 getRemoteAddr()
+            ip = request.getRemoteAddr();
+        }
+
+        // 如果 X-Forwarded-For 中有多个 IP 地址，返回第一个地址
+        if (ip != null && ip.contains(",")) {
+            ip = ip.split(",")[0];
+        }
+        return ip;
     }
 }